% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartdata.R
\docType{data}
\name{heartdata}
\alias{heartdata}
\title{Data from the Stanford Heart Transplant Center}
\format{
A numeric dataframe
}
\source{
the data is originally from the package named survival and transformed the data into a time-dependent covariate
}
\usage{
heartdata
}
\description{
Data from the Stanford Heart Transplant Center
}
\examples{
library(TRMST)
data(heartdata)
str(heartdata)
}
\keyword{datasets}
