% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_RMST.R
\name{T_RMST}
\alias{T_RMST}
\title{Title: the function of time-dependent RMST regression}
\usage{
T_RMST(data, time, tstop, id, status, event, fixed, varing)
}
\arguments{
\item{data}{Raw data set}

\item{time}{the name of the variable in data that indicates survival time}

\item{tstop}{the name of the variable in data that indicates the right  endpoint of the covariate observation time interval}

\item{id}{the name of the variable in data that identifies the different subjects}

\item{status}{the name of the variable in data that indicates  survival  status}

\item{event}{the name of the variable in date that indicates survival status in the covariate observation time interval}

\item{fixed}{the fixed covariates parameters (including the name of fixed covariates and reference group)}

\item{varing}{the time-dependent covariates parameters}
}
\value{
the result of time-dependent RMST
}
\description{
Title: the function of time-dependent RMST regression
}
\examples{
library(TRMST)
library(survival)
library(geepack)
data(heartdata)
fixed<-list(ZC=c("age1","age2","year","surgery"),ref=c(age1=0,age2=0,year=0,surgery=0))
varing<-list(ZC=c("transplant"),ref=c(transplant=0))
T_RMST(heartdata,time="time",tstop="tstop",id="id",status="status",event="event",fixed,varing)
}
