% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_RMST.R
\name{T_RMST}
\alias{T_RMST}
\title{Title: the function of time-dependent RMST regression}
\usage{
T_RMST(data,time,tstop,id,status,event,fixed,varing,tau)
}
\arguments{
\item{data}{Raw data set}

\item{time}{the name of the variable in data that indicates survival time}

\item{tstop}{the name of the variable in data that indicates the right  endpoint of the covariate observation time interval}

\item{id}{the name of the variable in data that identifies the different subjects}

\item{status}{the name of the variable in data that indicates  survival  status}

\item{event}{the name of the variable in date that indicates survival status in the covariate observation time interval}

\item{fixed}{the fixed covariates}

\item{varing}{the time-dependent covariates}

\item{tau}{the value to specify the truncation time point for the RMST,if it is null, indicates it is the largest follow-up time}
}
\value{
An object of type "geeglm"
}
\description{
Title: the function of time-dependent RMST regression
}
\examples{
library(TRMST)
data(heartdata)
fixed<-c("age1","age2","year","surgery")
varing<-c("transplant")
TRMST<-T_RMST(data=heartdata,time="time",tstop="tstop",id="id",status="status",event="event",fixed,varing,tau=NULL)
summary(TRMST)
}
\references{
Chengfeng Zhang,Baoyi Huang,Hongji Wu,Hao Yuan,Yawen Hou*,
Zheng Chen*.Restricted mean survival time regression model with time-dependent covariates.
Statistics in Medicine.2022;41(21):4081–4090.
}
